Added transitive adaptation: declaring an adapter "AB" from A to B, and "BC" from B to C, automatically implies adaptation from A to C via BC(AB()).  Used it to simplify the mechanism for adapting an IExecutable to an ICmdLineAppFactory.  Refactored some of the interface package's initial kludges to cleaner mechanisms based on adaptation.
Learned some interesting things along the way...  Classic classes don't
have __mro__, which made adapt() break on them.  The technique I was using
in the hopes of suppressing unwanted TypeErrors suppressed wanted ones, and
meanwhile there weren't any unwanted ones in PEAK.  Finally, Python 2.2
doesn't support 'super()' access to properties, and metamethods are
properties, so you have to use my new 'supermeta' to work around this when
you need to call a super() metamethod.  This is supposed to be fixed in
Python 2.3, but not backported 2.2 because Guido considers it a "feature".
Ah well.


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@1110 571e12c6-e1fa-0310-aee7-ff1267fa46bd