Fixed bugs found by Alexander Smishlajev's "Echo" class.  Specifically, I put back code to ensure that EigenRegistry and PropertyMap always use the "narrowest" interface possible, but it now uses a depth value to do so, which I find easier to understand than the old 'extends()' logic.  It also corresponds to the same logic used in the protocols package.  Anyway, the absence of this meant that ITwistedReactor was looked for even if IBasicReactor would suffice.  I put back the tests that checked this sort of thing, at least for EigenRegistry.
The second bug was that if a component with an "assembled" parent didn't
have any children already waiting for an assembly event, it didn't
request notification for events, and thus never notified any children that
subscribed later.  I fixed this by having components register with their
parents when their children register with them, if it is the first such
registration and the component would not have otherwise registered with
its parent.  Added a test case to cover this scenario.


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@1121 571e12c6-e1fa-0310-aee7-ff1267fa46bd