Added experimental 'invoke.c' script for POSIX-ish platforms with funky '#!' support, or lack thereof.  'invoke' is designed to be used like this:
   #!/usr/local/bin/invoke peak somearg otherarg...

This should work on most sane platforms with a long-enough commandline.
See http://homepages.cwi.nl/~aeb/std/hashexclam-1.html for details on the
insanely incompatible ways different Unixes interpret #! lines.  :(
Need to add a makefile for this and/or figure out how to get distutils to
build/install it on POSIX platforms.


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@1132 571e12c6-e1fa-0310-aee7-ff1267fa46bd