First phase of descriptor refactoring.  Once bindings now place a proxy for themselves in a class when activated in that class.  This avoids the need for Once-derivatives to "copy" themselves when they are found under more than one name (e.g in the same class or when they are created without a name initially).  Eventually this will let us do something similar with model.StructuralFeature objects.  Added hooks to the Once class family to allow an 'onSet()' method to be called that can alter a value being setattr'd.  This will let us do automatic 'adapt()' and 'setParentComponent()' when attributes are set or looked up.
Also, simplified the structure of the Pyrex OnceDescriptor base, using
Pyrex 0.7.2's ability to declare C-level attributes visible from Python.
(Previously, we used a custom descriptor to deal with this.)

NOTE: Error checking of bindings is partly BROKEN right now; bindings do
not verify that they haven't been placed in a class that doesn't support
them; they only verify that they have been told what name to use.  This
needs a little review, and the tutorial may need to be adjusted as well.


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@1185 571e12c6-e1fa-0310-aee7-ff1267fa46bd