Revised patch to allow <schema extends="url url url..."> in ZConfig schemas.  The 'extends=' attribute causes the named URLs to be used as a basis for the current schema.  The Keytype, datatype, and valuetype of the schema are inherited from the first listed schema that explicitly defines that parameter.
This is roughly equivalent to including the content of the extended schema
in the current schema's body.  As a result, you cannot override
any section types, sections, or keys defined by one of the base schemas.

Note that this patch is *still* highly experimental, and not supported by
the authors of ZConfig.


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@1398 571e12c6-e1fa-0310-aee7-ff1267fa46bd