Socket address schemes for tcp, udp, unix, and unix.dg (datagram). unix.dg is to udp as unix is to tcp.
These can currently be adapted to either IClientSocket or
IListeningSocket depending on what you want to do with them. Either way
gets you back a socket object that is configured as appropriate.

(Later, there will be and ISocketListener for use with reactors that
properly handles the "more than one socket for this address" issue, but
we can't really support that for fastcgi servers anyway right now)


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@1454 571e12c6-e1fa-0310-aee7-ff1267fa46bd