Partial implementation of the performance timer system described at:
  http://www.eby-sarna.com/pipermail/peak/2003-December/001020.html

It's missing the 'addListener()' function for the timer service, so you
can only add listeners directly to individual timers.  The actual
interfaces are also a bit different than what was originally proposed.
An exhaustive test suite is included, covering all of the implemented
behavior.  This draft implementation will probably get moved to Pyrex,
at least to create a a Timer base class with fast-path execution for
frequently called methods.  The goal is to have timer invocation be so
fast that leaving timers in production code isn't an issue.

(TODO: 'addListener()' for keys/wildcards, 'timer:' URL scheme, and
Pyrex port of Timer base.)


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@1562 571e12c6-e1fa-0310-aee7-ff1267fa46bd