Extensible cursor formatter support
Make the various cursor output formats defined by n2 generally
available, and make it possible to configure new ones.

It is now possible to call getFormatter(NAME, ...) on a cursor and get
an CursorFormatter (factory looked up in peak.cursor.formatters.NAME)
that is bound to it.  As they are very common, "header", "footer" (both
bool), and "delim" are available as keyword properties on all formatters
(and may be passed to getFormatter). They may be ignored by some
formatters. Other keyword args may be passed if you know the named
formatter accepts them, or you can set properties on the formatter that
it may bind to.

The resulting formatter may be called, passing a file-like object, to
format the cursor's data to the file.

For example, to export a subtree of LDAP in a LDIF file:

ldapcon = root.lookupComponent('ldap://...')
curs = ldapcon('o=fooblycorp', ldap.SCOPE_SUBTREE)
curs.getFormatter('ldif')(sys.stdout)


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@1566 571e12c6-e1fa-0310-aee7-ff1267fa46bd