Added 'ddt.web' command that runs DDT test documents in a browser.  Try it with::
  peak ddt.web http://fit.c2.com/files/HtmlSamples/Documents

and try running some tests.  This is still very crude; note that it doesn't
handle non-HTML documents well, and can't do its own directory indexes when
you use a local URL as the base.  It also lacks any error trapping, so you
have to keep an eye on the console you've run it from in order to find out
what's happening.  Nonetheless, it works pretty decently for such a rough
draft.


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@1642 571e12c6-e1fa-0310-aee7-ff1267fa46bd