Added '[Named Services]' section parser to 'peak.ini'.  This new section type functions almost identically to '[Component Factories]', except that the keys are property names rather than references to interfaces or other component keys.  This should be helpful for configuring shared services that all implement the same interface, such as SQL connection objects.
Note that to look up a named service, you simply use the appropriate
property name, as usual.  The difference is that you are guaranteed to
always use the same *instance* of the service, within a given service
area.

Also, added 'ddt.SQLChecker()', which verifies SQL results against a
test document.  'SQLChecker' makes use of the new "Named Services" feature
in order to ensure it's accessing the database you intend to test (as
opposed to say, the nearest SQL connection object).


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@1659 571e12c6-e1fa-0310-aee7-ff1267fa46bd