Minor 'peak.web' enhancements, to better support the "app as resource dirs" paradigm, as described in:
http://www.eby-sarna.com/pipermail/peak/2004-January/001166.html

Except that this is to support creating DDT document directories that are
web-publishable.  The main changes are to make resources traversable-to,
add a 'basename' attribute for filesystem resources (base filename, ala
Zope's 'id'), and add 'header' and 'footer' options to the 'list' DOMlet.

Working with this a bit has shown me that nameless security checks (i.e.
'interaction.allows(ob)') need a bit of work.  We should probably assume
that access to an object as a collection member is safe unless declared
otherwise, since the use case is to allow application objects to exclude
themselves from appearing in a collection.  But that'll need to change in
'peak.security', not 'peak.web'.

Also, traversal implementations are starting to proliferate, and probably
need a mini-framework added to pull them back together.


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@1684 571e12c6-e1fa-0310-aee7-ff1267fa46bd