Added 'peak.util.expr': tools for symbolic manipulation of Python expressions.  This first drop provides a mechanism for creating proxy types that invoke a builder function when expressions are executed, thus translating things like 'something.foo[23]' into something like::
    build("[]", build(".",something,"foo"), 23 )

Assuming, of course, that 'something' is a proxy instance.  Expression
proxies can wrap the node object types of your choice, so that you can
reuse this code for various types of query or other frameworks, without
having to code all the '__foo__' methods to do this sort of thing by hand.


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@1761 571e12c6-e1fa-0310-aee7-ff1267fa46bd