Added Zope 3-like "namespaces" to 'peak.web'.  Path segments in a URL may be prefixed with '"++some_id++"' in order to invoke a corresponding namespace handler registered under '"peak.web.namespaces.some_id". Namespace handlers must implement 'web.INamespaceHandler', and they are supplied with the original path segment as well as the separated namespace and name.  Also, as in Zope 3, '"@@foo"' is a shortcut for '"++view++foo"'. Builtin namespaces at this time include 'view', 'item', 'attr', 'skin', and 'resources'.  'skin' treats the rest of its path segment as a skin name, and sets the current skin, and 'resources' begins traversal to resources found in the current skin.  The other namespaces are as described at:
http://www.eby-sarna.com/pipermail/peak/2004-August/001712.html

Note, however, that this doesn't include the change to default traversal
described there, and there are a couple of other issues outstanding, like
defining permissions for views, and fixing legacy uses of "@@" in the
peak.web source.  We also don't have any registration mechanisms for easy
view declarations, or for setting up local "sites" or "sub-applications".

Also, this refactoring may break existing uses of 'web.traverseAttr', which
now has a completely different meaning and signature.


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@1813 571e12c6-e1fa-0310-aee7-ff1267fa46bd