Implemented 'with:content-type="whatever"', 'this:is="page"', 'with:fragment-layout="layoutpath"', etc., as described in:
  http://www.eby-sarna.com/pipermail/peak/2004-October/001835.html

Note that this means existing PWT pages will no longer function as web
pages unless you add 'with:page-layout="/default"' or 'this:is="page"' to
the top-level element.  You can use the shorter form if you don't need any
text outside the top-level element (such as whitespace, <!DOCTYPE, a DTD
fragment, or the <?xml PI).  Otherwise, you should use the longer
'with:page-layout="/default"' form, which means, "use the whole file as
a page.  The intended tradeoff is to favor page fragments over full pages,
pushing most HTML details into one or more page layouts.  But, it's not
clear just yet if this is the right thing to do, so this decision could
be reversed or tweaked later.


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@1907 571e12c6-e1fa-0310-aee7-ff1267fa46bd