Massive overhaul of indexing and inequality handling to linearize the speed of defining new methods.  Before this change, method definition time would rapidly increase as the number of tests applied to a given argument expression increased.  In some informal tests, defining 500 methods using unique inequalities (==1, ==2, etc.) used to take about 2 seconds.  Now, you can define 5000 such methods in the same amount of time, and the rate stays linear at about 2500 methods/second on my PC, including the time to parse and compile the expressions.
The primary downside to this change is a possible loss of stability or
functionality.  It passes all the tests, but it's quite possible that I'm
missing important tests.  It also makes the code even harder to understand
than it already was, and I'm not sure *I* understand it any more!  In fact,
I think I stopped understanding it a few hours before I finished this
change.  However, I hope that with a few cleanup refactoring passes I'll be
able to document and/or re-understand it, and maybe be more certain that it
is correct in all circumstances, not just the ones I've thought of so far
to test.

Oh, and speaking of such things, I also fixed Jay Parlar's other reported
bug: the weird "make_index" thing that I thought was an import circularity.
It's in fact to do with using 'dispatch.as()' while executing another
decorator.  Nested assignment advisors don't work so well; guess we should
move to Python 2.4 ASAP.


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@2052 571e12c6-e1fa-0310-aee7-ff1267fa46bd